Refactor: Directory change
